# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: EvenSteven - Collaborative expense settlement API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /host/auth:
    post:
      operationId: hostAuth
      tags: [host]
      summary: Exchange HOST_PASSWORD for a host token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostAuthInput"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostAuthResult"
        "401":
          description: Incorrect password
  /host/auth/check:
    get:
      operationId: checkHostAuth
      tags: [host]
      summary: Verify an existing host token
      parameters:
        - name: x-host-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token validity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostAuthCheck"

  # ── Houses ────────────────────────────────────────────────────────────────
  /houses:
    get:
      operationId: listHouses
      tags: [houses]
      summary: List all houses
      responses:
        "200":
          description: Houses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/House"
    post:
      operationId: createHouse
      tags: [houses]
      summary: Create a house (host only)
      parameters:
        - name: x-host-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseInput"
      responses:
        "201":
          description: House created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
  /houses/{id}:
    patch:
      operationId: updateHouse
      tags: [houses]
      summary: Update a house (host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-host-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseInput"
      responses:
        "200":
          description: House updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
        "404":
          description: House not found
    delete:
      operationId: deleteHouse
      tags: [houses]
      summary: Delete a house (host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-host-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: House not found

  # ── People ────────────────────────────────────────────────────────────────
...
  /events/{token}/expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses in an event
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Add an expense
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /events/{token}/expenses/{expenseId}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Settlements ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    HostAuthInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
    HostAuthResult:
      type: object
      required: [token]
      properties:
        token:
          type: string
    HostAuthCheck:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean

    # ── Houses ────────────────────────────────────────────────────────────
    House:
      type: object
      required: [id, name, crest, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        crest:
          type: string
        accentColor:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    HouseInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        crest:
          type: string
        accentColor:
          type: string
          nullable: true

    # ── People ────────────────────────────────────────────────────────────
...
    Expense:
      type: object
      required: [id, eventId, paidByMemberId, paidByMemberName, category, amount, splitType, participants, createdAt, updatedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        paidByMemberId:
          type: integer
        paidByMemberName:
          type: string
        category:
          type: string
          enum: [tickets, food, drinks, snacks, fuel, other]
        amount:
          type: integer
          description: Amount in paise
        description:
          type: string
          nullable: true
        splitType:
          type: string
          enum: [everyone, families, members]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseParticipant"
        createdByMemberId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ExpenseParticipant:
      type: object
      required: [memberId, memberName, shareAmount]
      properties:
        memberId:
          type: integer
        memberName:
          type: string
        shareAmount:
          type: integer
    ExpenseInput:
      type: object
      required: [paidByMemberId, category, amount, splitType, createdByMemberId]
      properties:
        paidByMemberId:
          type: integer
        category:
          type: string
          enum: [tickets, food, drinks, snacks, fuel, other]
        amount:
          type: integer
          minimum: 1
        description:
          type: string
          nullable: true
        splitType:
          type: string
          enum: [everyone, families, members]
        participantIds:
          type: array
          items:
            type: integer
          nullable: true
        familyIds:
          type: array
          items:
            type: integer
          nullable: true
        createdByMemberId:
          type: integer

    # ── Settlements ───────────────────────────────────────────────────────
...