# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EvenSteven - Collaborative expense settlement API
 * OpenAPI spec version: 0.2.0
 */
import type { ExpenseCategory } from './expenseCategory';
import type { ExpenseParticipant } from './expenseParticipant';
import type { ExpenseSplitType } from './expenseSplitType';

export interface Expense {
  id: number;
  eventId: number;
  paidByMemberId: number;
  paidByMemberName: string;
  category: ExpenseCategory;
  /** Amount in paise */
  amount: number;
  description?: string | null;
  splitType: ExpenseSplitType;
  participants: ExpenseParticipant[];
  createdByMemberId?: number;
  createdAt: Date;
  updatedAt: Date;
}