# lib/api-zod/src/generated/types/hostAuthInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EvenSteven - Collaborative expense settlement API
 * OpenAPI spec version: 0.2.0
 */

export interface HostAuthInput {
  /** @minLength 1 */
  password: string;
}