# lib/api-zod/src/generated/types/hostAuthResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EvenSteven - Collaborative expense settlement API
 * OpenAPI spec version: 0.2.0
 */

export interface HostAuthResult {
  token: string;
}